These can be int

+++ ESPGeiger/src/GMC/GMC.h
@@ -45,7 +45,7 @@ class GMC {
     AsyncHTTPRequest request;
   private:
     unsigned long lastPing = 0;
-    const unsigned long pingInterval = 1000 * 300;
+    const int pingInterval = 1000 * 300;
     static void httpRequestCb(void *optParm, AsyncHTTPRequest *request, int readyState);
 };
